scoring-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
tritonclient[http]==2.44.0
psycopg2-binary==2.9.9
numpy==1.26.4
prometheus-client==0.20.0
boto3==1.34.0
pyyaml==6.0.1
scipy==1.13.0
requests==2.31.0